requirements.txt
requests
gitpython
semver